requirements.txt
streamlit>=1.28.0
pyttsx3>=2.90
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0
webrtcvad>=2.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
plotly>=5.17.0
pandas>=2.0.0
sqlite3
wave
